NLP/AUTOMATED-RESUME-JOB-MATCHING-SYSTEM-USING-NATURAL-LANGUAGE-PROCESSING/backend/app/models.py
from pydantic import BaseModel
from typing import List

class JobDescription(BaseModel):
    title: str
    description: str

class MatchResponse(BaseModel):
    candidate_name: str
    match_score: float
    extracted_skills: List[str]
