NLP/AUTOMATED-RESUME-JOB-MATCHING-SYSTEM-USING-NATURAL-LANGUAGE-PROCESSING/backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
spacy
sentence-transformers
torch
numpy
requests
